polished some code, streamlined it
