+++ .vscode/settings.json
@@ -52,5 +52,8 @@
         "ProjectSettings/":true,
         "temp/":true,
         "Temp/":true
+    },
+    "files.associations": {
+        "iosfwd": "cpp"
     }
 }
