+++ Assets/Scripts/Generic/IShooter.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Events;
+
+public interface IShooter
+{
+    UnityEvent OnFinished { get; set; }
+
+    void setHolding(bool isHolding);
+}
