+++ Assets/Scripts/Generic/IShooter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 67aa6924070a330459b3e20286be5e2c
+guid: bc48e64531ee98544a81e1675d9590e9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
