+++ Assets/Scripts/Player/GunController.cs
@@ -3,7 +3,33 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-public class GunController : GenericGunController<GunData>
+public class GunController : InputUser
 {
+    IShooter shooter;
 
+
+    void Awake()
+    {
+        shooter = GetComponent<IShooter>();
+    }
+
+    void Start()
+    {
+
+    }
+
+    void OnDisable()
+    {
+
+    }
+    
+    void OnEnable()
+    {
+        input.attackEvent += shooter.setHolding;
+
+        shooter.OnFinished.AddListener(() =>
+        {
+            input.attackEvent -= shooter.setHolding;
+        });
+    }
 }
